# spec/04.api/api.yaml
openapi: 3.0.0

info:
  title: ECSite API
  description: ECSiteのWebAPI仕様。URL定義はケバブケース(-)を採用。
  version: 1.0.0

servers:
  - url: http://localhost:3001/dev/api/v1
    description: localhost
  - url: https://5adqe2wdk2.execute-api.ap-northeast-1.amazonaws.com/dev/api/v1
    description: dev server

tags:
  - name: meta
    description: メタ情報を扱うAPI。
  - name: goods
    description: 商品情報を扱うAPI。

paths:
  /health:
    get:
      tags:
        - meta
      summary: 疎通確認をする。
      description: 疎通確認をする。
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: string
                example: Hello World!!

  /generate-presigned-url:
    get:
      tags:
        - meta
      summary: 画像アップロード用の一時URLを取得する。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: (TBD:チケットNoのECSITE-14で対応) 認証情報。
          schema:
            type: string
        - name: name
          in: path
          required: true
          description: 画像のファイル名称
          schema:
            type: string
        - name: type
          in: path
          required: true
          description: 画像のコンテンツタイプ
          schema:
            type: string
      description: 画像アップロード用の一時URL(AWS S3 Presigned URL)を取得する。
      operationId: generatePresignedUrl
      responses:
        "200":
          description: 画像アップロード用の一時URL(AWS S3 Presigned URL)の生成に成功。
          content:
            application/json:
              schema:
                type: object
                properties:
                  presignedUrl:
                    type: string
        "400":
          description: Invalid input
        "500":
          description: Server error

  /goods:
    get:
      tags:
        - goods
      summary: 販売者が所有する商品の一覧を取得する。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: (TBD:チケットNoのECSITE-14で対応) 認証情報。
          schema:
            type: string
      responses:
        "200":
          description: 販売者が所有する商品の一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goods"
              examples:
                goodsList:
                  $ref: "#/components/examples/GoodsList"
  /goods/new:
    post:
      tags:
        - goods
      summary: 商品を新規登録する。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: (TBD:チケットNoのECSITE-14で対応) 認証情報。
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Goods"
            examples:
              goods:
                $ref: "#/components/examples/Goods"
      responses:
        "201":
          description: Created

  /goods/{id}:
    get:
      tags:
        - goods
      summary: 指定された商品IDの情報を取得する。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: (TBD:チケットNoのECSITE-14で対応) 認証情報。
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: 商品ID
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goods"
              examples:
                goods:
                  $ref: "#/components/examples/Goods"
    put:
      tags:
        - goods
      summary: 指定された商品IDの情報を更新する。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: (TBD:チケットNoのECSITE-14で対応) 認証情報。
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: 商品ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Goods"
            examples:
              goods:
                $ref: "#/components/examples/Goods"
      responses:
        "200":
          description: Successful update
    delete:
      tags:
        - goods
      summary: 指定された商品IDの情報を削除する。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: (TBD:チケットNoのECSITE-14で対応) 認証情報。
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: 商品ID
          schema:
            type: integer
      responses:
        "204":
          description: Successfully deleted

components:
  schemas:
    Goods:
      type: object
      properties:
        id:
          type: integer
          description: 商品ID。各商品を一意に識別するための番号やコード。
        name:
          type: string
          description: 商品名。ユーザーが商品を検索や認識するための名前。
        explanation:
          type: string
          description: 商品説明。商品の詳細な情報や特徴を説明するテキスト。
        price:
          type: integer
          description: 商品の販売価格。
        image:
          type: string
          description: 商品の実際の見た目を示す画像(写真やイラスト)が格納されたS3のURL。
        category:
          type: string
          description: 商品の分類。例：「家電」「ファッション」「食品」など。
        createAt:
          type: string
          description: 商品情報の作成日時。
        updateAt:
          type: string
          description: 商品情報の更新日時。
      required:
        - id
        - name
        - price
  examples:
    GoodsList:
      value:
        - id: 1
          name: けろけろ
        - id: 2
          name: ドドドドドドドドド
    Goods:
      value:
        id: 1
        name: けろけろ
        explanation: けろ？けろけろけろ・・・けろ！
        price: 2980
        image: s3://{}
        category: カエル
        createAt: 20230828T050442+0900
        updateAt: 20230828T050442+0900